Integrate show_for method for User Meta Container
This is needed, when we are allowing access for custom user roles to the administration interface.

Custom User Roles may be asked to fill in user details, like address, city, state etc. This change will allow the container to be visible for them.

The method `show_for` accepts an array, containing either or both `roles` and `capabilities`, which will be limiting the container visibility with logical `AND`, so when passing both, both will need to be evaluated to `true`.