Update Container.php
The library has a beautiful notification system, but there is no mechanism to add notifications from outside of the class if needed. The add_notification function will give this possibility. I have tested that on a theme_options container, not sure what effect notifications have on other container types. Maybe the same can be done for error messages.